backend/requirements.txt
# FastAPI & server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0

# Meta / Facebook
facebook-business>=19.0.0
httpx>=0.26.0

# Claude
anthropic>=0.18.0

# Config & security
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
cryptography>=42.0.0

# Scheduling
apscheduler>=3.10.4

# Utils
python-dateutil>=2.8.2
numpy>=1.26.0
